0.7.4 - Flattened UI, fixed logic issue
Returned to this to make it more appealing just before I begin working on blueprinting and constructing the cabinet itself. I'm now much happier with this in it's current state - expect tweaks to it's looks soon (especially closer to EP marketplace).

+++ app/games.js
@@ -7,39 +7,30 @@ const processWindows = require("node-process-windows");
 
 //Starts emulator for given game
 function openGame(gameConsole, game) {
-	if(emulatorProc != null) {
-		return;
-	}
+	if(emulatorProc != null) return;
 
 	//Get paths
 	var emulatorPath = Files.getEmulatorPath(gameConsole);
 	var gamePath = Files.getRomPath(gameConsole, game);
-	if(gamePath == null || emulatorPath == null) {
-		return;
-	}
+	if(gamePath == null || emulatorPath == null) return;
 
 	//Read in emu config to be used in childProc
 	var config = Files.getConfig(gameConsole, game);
 	console.log(config);
 
 	emulatorProc = child.execFile(emulatorPath, [gamePath].concat(config.cliArgs), (error, stdout, stderr) => {
-	  if (error) {
-		throw error;
-	  }
+		if (error) throw error;
 	});
 
 	//Wait configured time
 	window.setTimeout(function() {
 		app.remote.getCurrentWindow().setAlwaysOnTop(false);
-	},config.waitTime);
+	}, config.waitTime);
 }
 
 //Listener for the press of the kill button which will kill the proc in its tracks
 ipcRenderer.on("childProc", (event, arg) => {
-	console.log(arg);
-	if(emulatorProc != null) {
-		emulatorProc.kill();
-	}
+	if(emulatorProc != null) emulatorProc.kill();
 	//Refocus on close
 	emulatorProc = null;
 	processWindows.focusWindow("Annode");
